third_party/blink/web_tests/external/wpt/preload/prefetch-cache-expected.txt

This is a testharness.js-based test.
[FAIL] Prefetch should respect cache-control: no-cache
  assert_equals: expected 2 but got 1
[FAIL] Prefetch should respect cache-control: max-age=0
  assert_equals: expected 2 but got 1
Harness: the test ran to completion.
